+++ Source/ProjectFirma.Web/ProjectFirma.Web.csproj
@@ -321,7 +321,6 @@
     <Compile Include="Views\Agreement\AgreementDetail.cs" />
     <Compile Include="Views\Agreement\AgreementDetailViewData.cs" />
     <Compile Include="Views\ExcelUpload\ManageFbmsUploadViewModel.cs" />
-    <Compile Include="Views\ExcelUpload\InvoiceStageImports.cs" />
     <Compile Include="Views\ExcelUpload\BudgetStageImportsHelper.cs" />
     <Compile Include="Views\ExcelUpload\ImportEtlExcelFileViewModel.cs" />
     <Compile Include="Views\ExcelUpload\ImportEtlExcelFileViewData.cs" />
