+++ Source/ProjectFirmaModels/Models/ExcelUpload/BudgetStageImport.cs
@@ -1,5 +1,5 @@
 ﻿/*-----------------------------------------------------------------------
-<copyright file="ImportGdbFile.cs" company="Tahoe Regional Planning Agency and Sitka Technology Group">
+<copyright file="BudgetStageImport.cs" company="Tahoe Regional Planning Agency and Sitka Technology Group">
 Copyright (c) Tahoe Regional Planning Agency and Sitka Technology Group. All rights reserved.
 <author>Sitka Technology Group</author>
 </copyright>
@@ -23,31 +23,9 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
-using LtInfo.Common;
-using Microsoft.Web.Infrastructure.DynamicValidationHelper;
-using ProjectFirmaModels.Models.ExcelUpload;
 
-
-namespace ProjectFirmaModels.Models
+namespace ProjectFirmaModels.Models.ExcelUpload
 {
-    /// <summary>
-    /// An exception generated when a cell value is bad
-    /// </summary>
-    public class BudgetStageImportBadCellException : SitkaDisplayErrorException
-    {
-        public string ColumnName;
-        public int RowIndex;
-        public string CellValue;
-
-        public BudgetStageImportBadCellException(string columnName, int rowIndex, string cellValue, string errorMessage, Exception innerException) : base(string.Format("{0} - Cell {1}{2} - Cell Value \"{3}\"", errorMessage, columnName, rowIndex + 1, cellValue), innerException)
-        {
-            ColumnName = columnName;
-            CellValue = cellValue;
-            RowIndex = rowIndex;
-        }
-    }
-
-   
     public class BudgetStageImport
     {
         public readonly string BusinessAreaKey;
@@ -72,15 +50,6 @@ namespace ProjectFirmaModels.Models
         {
             var rowIndex = keyValuePair.Key;
             var dr = keyValuePair.Value;
-
-            /*
-             *
-             *
-             *
-             */
-
-            // Are you there tom?
-
 
             //var columnNames =  BudgetStageImports.GetBudgetColumnLetterToColumnNameDictionary();
             var columnNameToLetterDict = BudgetStageImports.GetBudgetColumnNameToColumnLetterDictionary();
